Autoray as a backend for qml.math (#1202)
* Use autoray as a backend

* remove tensorbox

* fixes

* push

* push

* linting

* test fix

* linting

* fix

* fix

* fix

* reorginzation

* fix

* linting

* fix docs

* fix docs

* attempt to increase coverage

* fix

* black tests

* merge master

* Update pennylane/math/single_dispatch.py

* Update pennylane/math/multi_dispatch.py


* Update pennylane/math/multi_dispatch.py


* suggested changes

* revert

* Update pennylane/math/single_dispatch.py

* Update pennylane/math/multi_dispatch.py


* Apply suggestions from code review


* suggested changes

* increase coverage

* increase coverage

* increase coverage

* Update tests/math/test_functions.py


* add docstrings

* update dev guide
