+++ .github/workflows/tests.yml
@@ -19,8 +19,8 @@ jobs:
     strategy:
       matrix:
         config:
-          - {python-version: 3.7, interfaces: ['tf']}
           - {python-version: 3.7, interfaces: ['torch']}
+          - {python-version: 3.7, interfaces: ['tf']}
           - {python-version: 3.8, interfaces: ['tf', 'torch', 'jax']}
           - {python-version: 3.9, interfaces: []}
