+++ requirements.txt
@@ -7,3 +7,4 @@ toml
 appdirs
 semantic_version==2.6
 dask[delayed]
+autoray
