+++ setup.py
@@ -25,6 +25,7 @@ requirements = [
     "toml",
     "appdirs",
     "semantic_version==2.6",
+    "autoray"
 ]
 
 info = {
