+++ tests/math/test_functions.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Unit tests for the TensorBox functional API in pennylane.math.fn
+"""Unit tests for the TensorBox functional API in pennylane.fn.fn
 """
 import itertools
 import numpy as onp
@@ -19,7 +19,7 @@ import pytest
 
 import pennylane as qml
 from pennylane import numpy as np
-from pennylane.math import fn
+from pennylane import math as fn
 
 
 tf = pytest.importorskip("tensorflow", minversion="2.1")
@@ -29,7 +29,7 @@ jnp = pytest.importorskip("jax.numpy")
 
 
 class TestGetMultiTensorbox:
-    """Tests for the _get_multi_tensorbox utility function"""
+    """Tests for the _multi_dispatch utility function"""
 
     def test_exception_tensorflow_and_torch(self):
         """Test that an exception is raised if the sequence of tensors contains
@@ -39,7 +39,7 @@ class TestGetMultiTensorbox:
         z = torch.tensor([0.6])
 
         with pytest.raises(ValueError, match="Tensors contain mixed types"):
-            fn._get_multi_tensorbox([x, y, z])
+            fn._multi_dispatch([x, y, z])
 
     def test_warning_tensorflow_and_autograd(self):
         """Test that a warning is raised if the sequence of tensors contains
@@ -48,7 +48,7 @@ class TestGetMultiTensorbox:
         y = np.array([0.5, 0.1])
 
         with pytest.warns(UserWarning, match="Consider replacing Autograd with vanilla NumPy"):
-            fn._get_multi_tensorbox([x, y])
+            fn._multi_dispatch([x, y])
 
     def test_warning_torch_and_autograd(self):
         """Test that a warning is raised if the sequence of tensors contains
@@ -57,39 +57,39 @@ class TestGetMultiTensorbox:
         y = np.array([0.5, 0.1])
 
         with pytest.warns(UserWarning, match="Consider replacing Autograd with vanilla NumPy"):
-            fn._get_multi_tensorbox([x, y])
+            fn._multi_dispatch([x, y])
 
     def test_return_tensorflow_box(self):
         """Test that TensorFlow is correctly identified as the dispatching library."""
         x = tf.Variable([1.0, 2.0, 3.0])
         y = onp.array([0.5, 0.1])
 
-        res = fn._get_multi_tensorbox([y, x])
-        assert res.interface == "tf"
+        res = fn._multi_dispatch([y, x])
+        assert res == "tensorflow"
 
     def test_return_torch_box(self):
         """Test that Torch is correctly identified as the dispatching library."""
         x = torch.tensor([1.0, 2.0, 3.0])
         y = onp.array([0.5, 0.1])
 
-        res = fn._get_multi_tensorbox([y, x])
-        assert res.interface == "torch"
+        res = fn._multi_dispatch([y, x])
+        assert res == "torch"
 
     def test_return_autograd_box(self):
         """Test that autograd is correctly identified as the dispatching library."""
         x = np.array([1.0, 2.0, 3.0])
         y = [0.5, 0.1]
 
-        res = fn._get_multi_tensorbox([y, x])
-        assert res.interface == "autograd"
+        res = fn._multi_dispatch([y, x])
+        assert res == "autograd"
 
     def test_return_numpy_box(self):
         """Test that NumPy is correctly identified as the dispatching library."""
         x = onp.array([1.0, 2.0, 3.0])
         y = [0.5, 0.1]
 
-        res = fn._get_multi_tensorbox([y, x])
-        assert res.interface == "numpy"
+        res = fn._multi_dispatch([y, x])
+        assert res == "numpy"
 
 
 test_abs_data = [
@@ -107,7 +107,7 @@ test_abs_data = [
 def test_abs(t):
     """Test that the absolute function works for a variety
     of input"""
-    res = fn.abs_(t)
+    res = fn.abs(t)
     assert fn.allequal(res, [1, 2, 5])
 
 
@@ -453,7 +453,7 @@ class TestDot:
         [tf.constant([[1, 2], [3, 4]]), onp.array([6, 7])],
         [tf.Variable([[1, 2], [3, 4]]), tf.Variable([6, 7])],
         [jnp.array([[1, 2], [3, 4]]), jnp.array([6, 7])],
-        [np.array([[1, 2], [3, 4]]), jnp.array([6, 7])],
+        [onp.array([[1, 2], [3, 4]]), jnp.array([6, 7])],
     ]
 
     @pytest.mark.parametrize("t1, t2", matrix_vector_product_data)
@@ -576,8 +576,8 @@ interface_test_data = [
     [onp.array([1, 2, 3]), "numpy"],
     [np.array([1, 2, 3]), "autograd"],
     [torch.tensor([1, 2, 3]), "torch"],
-    [tf.Variable([1, 2, 3]), "tf"],
-    [tf.constant([1, 2, 3]), "tf"],
+    [tf.Variable([1, 2, 3]), "tensorflow"],
+    [tf.constant([1, 2, 3]), "tensorflow"],
     [jnp.array([1, 2, 3]), "jax"],
 ]
 
@@ -596,6 +596,15 @@ def test_toarray(t):
     """Test that the toarray method correctly converts the input
     tensor into a NumPy array."""
     res = fn.toarray(t)
+    assert fn.allequal(res, t)
+    assert isinstance(res, onp.ndarray)
+
+
+@pytest.mark.parametrize("t", test_data)
+def test_numpy(t):
+    """Test that the to_numpy method correctly converts the input
+    tensor into a NumPy array."""
+    res = fn.to_numpy(t)
     assert fn.allequal(res, t)
     assert isinstance(res, onp.ndarray)
 
@@ -691,6 +700,11 @@ class TestRequiresGrad:
             tape.watch([t1, t2])
             assert fn.requires_grad(t1)
             assert fn.requires_grad(t2)
+
+    def test_unknown_interface(self):
+        """Test that an error is raised if the interface is unknown"""
+        with pytest.raises(ValueError, match="unknown object"):
+            fn.requires_grad(type("hello", tuple(), {})())
 
 
 shape_test_data = [
@@ -793,28 +807,28 @@ class TestSum:
     def test_array(self):
         """Test that sum, called without the axis arguments, returns a scalar"""
         t = np.array([[0.1, 0.2, 0.3], [0.4, 0.5, 0.6]])
-        res = fn.sum_(t)
+        res = fn.sum(t)
         assert isinstance(res, np.ndarray)
         assert fn.allclose(res, 2.1)
 
     def test_tensorflow(self):
         """Test that sum, called without the axis arguments, returns a scalar"""
         t = tf.Variable([[0.1, 0.2, 0.3], [0.4, 0.5, 0.6]])
-        res = fn.sum_(t)
+        res = fn.sum(t)
         assert isinstance(res, tf.Tensor)
         assert fn.allclose(res, 2.1)
 
     def test_torch(self):
         """Test that sum, called without the axis arguments, returns a scalar"""
         t = torch.tensor([[0.1, 0.2, 0.3], [0.4, 0.5, 0.6]])
-        res = fn.sum_(t)
+        res = fn.sum(t)
         assert isinstance(res, torch.Tensor)
         assert fn.allclose(res, 2.1)
 
     def test_jax(self):
         """Test that sum, called without the axis arguments, returns a scalar"""
         t = jnp.array([[0.1, 0.2, 0.3], [0.4, 0.5, 0.6]])
-        res = fn.sum_(t)
+        res = fn.sum(t)
         assert fn.allclose(res, 2.1)
 
     @pytest.mark.parametrize(
@@ -829,7 +843,7 @@ class TestSum:
     def test_sum_axis(self, t1):
         """Test that passing the axis argument allows for summing along
         a specific axis"""
-        res = fn.sum_(t1, axis=(0, 2))
+        res = fn.sum(t1, axis=(0, 2))
 
         # if tensorflow or pytorch, extract view of underlying data
         if hasattr(res, "numpy"):
@@ -850,7 +864,7 @@ class TestSum:
     def test_sum_axis_keepdims(self, t1):
         """Test that passing the axis argument allows for summing along
         a specific axis, while keepdims avoids the summed dimensions from being removed"""
-        res = fn.sum_(t1, axis=(0, 2), keepdims=True)
+        res = fn.sum(t1, axis=(0, 2), keepdims=True)
 
         # if tensorflow or pytorch, extract view of underlying data
         if hasattr(res, "numpy"):
@@ -926,6 +940,17 @@ class TestTake:
         assert fn.allclose(res, expected)
 
     @pytest.mark.parametrize("t", take_data)
+    def test_array_indexing_along_axis(self, t):
+        """Test that indexing with a sequence properly extracts
+        the elements from the specified tensor axis"""
+        indices = [0, 1, -2]
+        res = fn.take(t, indices, axis=2)
+        expected = np.array(
+            [[[1, 2, 1], [3, 4, 3], [-1, 1, -1]], [[5, 6, 5], [0, -1, 0], [2, 1, 2]]]
+        )
+        assert fn.allclose(res, expected)
+
+    @pytest.mark.parametrize("t", take_data)
     def test_multidimensional_indexing_along_axis(self, t):
         """Test that indexing with a sequence properly extracts
         the elements from the specified tensor axis"""
@@ -935,6 +960,25 @@ class TestTake:
             [[[[1, 2], [1, 2]], [[3, 4], [1, 2]]], [[[5, 6], [5, 6]], [[0, -1], [5, 6]]]]
         )
         assert fn.allclose(res, expected)
+
+    def test_multidimensional_indexing_along_axis_autograd(self):
+        """Test that indexing with a sequence properly extracts
+        the elements from the specified tensor axis"""
+        t = np.array([[[1, 2], [3, 4], [-1, 1]], [[5, 6], [0, -1], [2, 1]]])
+        indices = np.array([[0, 0], [1, 0]])
+
+        def cost_fn(t):
+            return fn.sum(fn.take(t, indices, axis=1))
+
+        res = cost_fn(t)
+        expected = np.sum(
+            np.array([[[[1, 2], [1, 2]], [[3, 4], [1, 2]]], [[[5, 6], [5, 6]], [[0, -1], [5, 6]]]])
+        )
+        assert fn.allclose(res, expected)
+
+        grad = qml.grad(cost_fn)(t)
+        expected = np.array([[[3, 3], [1, 1], [0, 0]], [[3, 3], [1, 1], [0, 0]]])
+        assert fn.allclose(grad, expected)
 
 
 where_data = [
@@ -1043,7 +1087,11 @@ class TestDiag:
 
     @pytest.mark.parametrize(
         "a, interface",
-        [[np.array(0.5), "autograd"], [tf.Variable(0.5), "tf"], [torch.tensor(0.5), "torch"]],
+        [
+            [np.array(0.5), "autograd"],
+            [tf.Variable(0.5), "tensorflow"],
+            [torch.tensor(0.5), "torch"],
+        ],
     )
     def test_sequence(self, a, interface):
         """Test that a sequence is automatically converted into
@@ -1322,3 +1370,25 @@ def test_gather(tensor):
     res = fn.gather(tensor, indices)
     expected = np.array([[-1, -6, -3], [1, 2, 3]])
     assert fn.allclose(res, expected)
+
+
+class TestCoercion:
+    """Test that TensorFlow and PyTorch correctly coerce types"""
+
+    def test_tensorflow_coercion(self):
+        """Test tensorflow coercion"""
+        tensors = [tf.Variable([0.2]), np.array([1, 2, 3]), tf.constant(1 + 3j, dtype=tf.complex64)]
+        res = qml.math.coerce(tensors, like="tensorflow")
+        dtypes = [r.dtype for r in res]
+        assert all(d is tf.complex64 for d in dtypes)
+
+    def test_torch_coercion(self):
+        """Test Torch coercion"""
+        tensors = [
+            torch.tensor([0.2]),
+            np.array([1, 2, 3]),
+            torch.tensor(1 + 3j, dtype=torch.complex64),
+        ]
+        res = qml.math.coerce(tensors, like="torch")
+        dtypes = [r.dtype for r in res]
+        assert all(d is torch.complex64 for d in dtypes)
